broke up threads/messages/runs routers and tests
