+++ tests/integration/api/test_threads.py
@@ -1,16 +1,14 @@
 """Test the API endpoints for assistants."""
 
 import os
-
 import pytest
 from fastapi import HTTPException, status
 from fastapi.testclient import TestClient
 from openai.types.beta import Thread, ThreadDeleted
 from openai.types.beta.thread import ToolResourcesCodeInterpreter, ToolResources
-from openai.types.beta.threads import TextContentBlock, Text, Message, MessageDeleted
+from openai.types.beta.threads import TextContentBlock, Text
 from leapfrogai_api.backend.types import (
     ModifyThreadRequest,
-    ModifyMessageRequest,
 )
 from leapfrogai_api.routers.openai.requests.create_message_request import (
     CreateMessageRequest,
@@ -18,8 +16,7 @@ from leapfrogai_api.routers.openai.requests.create_message_request import (
 from leapfrogai_api.routers.openai.requests.create_thread_request import (
     CreateThreadRequest,
 )
-from leapfrogai_api.routers.openai.threads import router as threads_router
-from leapfrogai_api.routers.openai.files import router as files_router
+from leapfrogai_api.routers.openai.threads import router
 
 
 class MissingEnvironmentVariable(Exception):
@@ -36,13 +33,16 @@ except KeyError as exc:
         "Please check the api README for instructions on obtaining this token."
     ) from exc
 
-threads_client = TestClient(threads_router, headers=headers)
-files_client = TestClient(files_router, headers=headers)
+
+@pytest.fixture(scope="session")
+def app_client():
+    with TestClient(router, headers=headers) as client:
+        yield client
 
 
 # Create a thread with the previously created file and fake embeddings
 @pytest.fixture(scope="session")
-def create_thread():
+def create_thread(app_client):
     """Create a thread for testing. Requires a running Supabase instance."""
 
     request = CreateThreadRequest(
@@ -51,11 +51,11 @@ def create_thread():
         metadata={},
     )
 
-    return threads_client.post("/openai/v1/threads", json=request.model_dump())
+    return app_client.post("/openai/v1/threads", json=request.model_dump())
 
 
 @pytest.fixture(scope="session")
-def create_message(create_thread):
+def create_message(app_client, create_thread):
     """Create a message for testing. Requires a running Supabase instance."""
 
     request = CreateMessageRequest(
@@ -67,7 +67,7 @@ def create_message(create_thread):
         metadata={},
     )
 
-    message = threads_client.post(
+    message = app_client.post(
         f"/openai/v1/threads/{create_thread.json()['id']}/messages",
         json=request.model_dump(),
     )
@@ -75,7 +75,7 @@ def create_message(create_thread):
     return {"message": message, "thread_id": create_thread.json()["id"]}
 
 
-def test_code_interpreter_fails():
+def test_code_interpreter_fails(app_client):
     """Test code interpreter fails."""
 
     tool_resources = ToolResources(
@@ -88,11 +88,11 @@ def test_code_interpreter_fails():
         metadata={},
     )
 
-    with pytest.raises(HTTPException) as exc:
-        threads_client.post("/openai/v1/threads", json=request.model_dump())
+    with pytest.raises(HTTPException) as code_exc:
+        app_client.post("/openai/v1/threads", json=request.model_dump())
 
-    assert exc.value.status_code == status.HTTP_400_BAD_REQUEST
-    assert exc.value.detail == f"Unsupported tool resource: {tool_resources}"
+    assert code_exc.value.status_code == status.HTTP_400_BAD_REQUEST
+    assert code_exc.value.detail == f"Unsupported tool resource: {tool_resources}"
 
 
 def test_create_thread(create_thread):
@@ -101,49 +101,17 @@ def test_create_thread(create_thread):
     assert Thread.model_validate(create_thread.json()), "Create should create a Thread."
 
 
-def test_create_message(create_message):
-    """Test creating a message. Requires a running Supabase instance."""
-    assert create_message["message"].status_code == status.HTTP_200_OK
-    assert Message.model_validate(
-        create_message["message"].json()
-    ), "Create should create a Message."
-
-
-def test_get_thread(create_thread):
+def test_get_thread(app_client, create_thread):
     """Test getting a threads. Requires a running Supabase instance."""
     threads_id = create_thread.json()["id"]
-    get_response = threads_client.get(f"/openai/v1/threads/{threads_id}")
+    get_response = app_client.get(f"/openai/v1/threads/{threads_id}")
     assert get_response.status_code == status.HTTP_200_OK
     assert Thread.model_validate(
         get_response.json()
     ), f"Get should return Thread {threads_id}."
 
 
-def test_get_message(create_message):
-    """Test getting a messages. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    get_response = threads_client.get(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}"
-    )
-    assert get_response.status_code == status.HTTP_200_OK
-    assert Message.model_validate(
-        get_response.json()
-    ), f"Get should return Message {message_id}."
-
-
-def test_list_message(create_message):
-    """Test listing messages. Requires a running Supabase instance."""
-    thread_id = create_message["thread_id"]
-    list_response = threads_client.get(f"/openai/v1/threads/{thread_id}/messages")
-    assert list_response.status_code == status.HTTP_200_OK
-    for message_object in list_response.json()["data"]:
-        assert Message.model_validate(
-            message_object
-        ), "Should return a list of Message."
-
-
-def test_modify_thread(create_thread):
+def test_modify_thread(app_client, create_thread):
     """Test modifying a thread. Requires a running Supabase instance."""
     thread_id = create_thread.json()["id"]
     request = ModifyThreadRequest(
@@ -151,7 +119,7 @@ def test_modify_thread(create_thread):
         metadata={"test": "modified"},
     )
 
-    modify_response = threads_client.post(
+    modify_response = app_client.post(
         f"/openai/v1/threads/{thread_id}",
         json=request.model_dump(),
     )
@@ -162,72 +130,10 @@ def test_modify_thread(create_thread):
     ), "Should be modified."
 
 
-def test_modify_message(create_message):
-    """Test modifying a thread. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    request = ModifyMessageRequest(
-        metadata={"test": "modified"},
-    )
-
-    modify_response = threads_client.post(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}",
-        json=request.model_dump(),
-    )
-    assert modify_response.status_code == status.HTTP_200_OK
-    assert Message.model_validate(modify_response.json()), "Should return a Message."
-    assert (
-        modify_response.json()["metadata"]["test"] == "modified"
-    ), "Should be modified."
-
-
-def test_get_modified_thread(create_thread):
-    """Test getting a modified threads. Requires a running Supabase instance."""
-    threads_id = create_thread.json()["id"]
-    get_modified_response = threads_client.get(f"/openai/v1/threads/{threads_id}")
-    assert get_modified_response.status_code == status.HTTP_200_OK
-    assert Thread.model_validate(
-        get_modified_response.json()
-    ), f"Get should return modified Thread {threads_id}."
-    assert (
-        get_modified_response.json()["metadata"]["test"] == "modified"
-    ), "Should be modified."
-
-
-def test_get_modified_message(create_message):
-    """Test getting a modified threads. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    get_modified_response = threads_client.get(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}"
-    )
-    assert get_modified_response.status_code == status.HTTP_200_OK
-    assert Message.model_validate(
-        get_modified_response.json()
-    ), f"Get should return modified Message {message_id}."
-    assert (
-        get_modified_response.json()["metadata"]["test"] == "modified"
-    ), "Should be modified."
-
-
-def test_delete_message(create_message):
-    """Test deleting a message. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    delete_response = threads_client.delete(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}"
-    )
-    assert delete_response.status_code == status.HTTP_200_OK
-    assert MessageDeleted.model_validate(
-        delete_response.json()
-    ), "Should return a MessageDeleted object."
-    assert delete_response.json()["deleted"] is True, "Should be able to delete."
-
-
-def test_delete_thread(create_thread):
+def test_delete_thread(app_client, create_thread):
     """Test deleting a thread. Requires a running Supabase instance."""
     thread_id = create_thread.json()["id"]
-    delete_response = threads_client.delete(f"/openai/v1/threads/{thread_id}")
+    delete_response = app_client.delete(f"/openai/v1/threads/{thread_id}")
     assert delete_response.status_code == status.HTTP_200_OK
     assert ThreadDeleted.model_validate(
         delete_response.json()
@@ -235,54 +141,25 @@ def test_delete_thread(create_thread):
     assert delete_response.json()["deleted"] is True, "Should be able to delete."
 
 
-def test_delete_twice_thread(create_thread):
+def test_delete_twice_thread(app_client, create_thread):
     """Test deleting a thread twice. Requires a running Supabase instance."""
     thread_id = create_thread.json()["id"]
-    delete_response = threads_client.delete(f"/openai/v1/threads/{thread_id}")
+    delete_response = app_client.delete(f"/openai/v1/threads/{thread_id}")
     assert delete_response.status_code == status.HTTP_200_OK
     assert ThreadDeleted.model_validate(
         delete_response.json()
     ), "Should return a ThreadDeleted object."
-    assert (
-        delete_response.json()["deleted"] is False
-    ), "Should not be able to delete twice."
-
-
-def test_delete_twice_message(create_message):
-    """Test deleting a message twice. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    delete_response = threads_client.delete(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}"
-    )
-    assert delete_response.status_code == status.HTTP_200_OK
-    assert MessageDeleted.model_validate(
-        delete_response.json()
-    ), "Should return a MessageDeleted object."
     assert (
         delete_response.json()["deleted"] is False
     ), "Should not be able to delete twice."
 
 
 @pytest.mark.xfail
-def test_get_nonexistent_thread(create_thread):
+def test_get_nonexistent_thread(app_client, create_thread):
     """Test getting a nonexistent thread. Requires a running Supabase instance."""
     thread_id = create_thread.json()["id"]
-    fail_response = threads_client.get(f"/openai/v1/threads/{thread_id}")
+    fail_response = app_client.get(f"/openai/v1/threads/{thread_id}")
     assert fail_response.status_code == status.HTTP_404_NOT_FOUND
     assert (
         fail_response.json().get("detail") == "Thread not found"
     ), f"Get should not return deleted Thread {thread_id}."
-
-
-def test_get_nonexistent_message(create_message):
-    """Test getting a nonexistent message. Requires a running Supabase instance."""
-    message_id = create_message["message"].json()["id"]
-    thread_id = create_message["thread_id"]
-    get_response = threads_client.get(
-        f"/openai/v1/threads/{thread_id}/messages/{message_id}"
-    )
-    assert get_response.status_code == status.HTTP_200_OK
-    assert (
-        get_response.json() is None
-    ), f"Get should not return deleted Message {message_id}."
